+++ packages/node-core/src/index.ts
@@ -43,7 +43,6 @@ export { initializeEsmLoader } from './sdk/esmLoader';
 export { isCjs } from './utils/detection';
 export { ensureIsWrapped } from './utils/ensureIsWrapped';
 export { createMissingInstrumentationContext } from './utils/createMissingInstrumentationContext';
-export { envToBool } from './utils/envToBool';
 export { makeNodeTransport, type NodeTransportOptions } from './transports';
 export type { HTTPModuleRequestIncomingMessage } from './transports/http-module';
 export { NodeClient } from './sdk/client';